refactor(server): rewrite server without framework
So i started writing series on CRA+SSR here https://vfeskov.com/tags/CRA-with-SSR-series/ , and I didn't use any backend framework there and I liked it. I wanted to put a link to Win A Beer in the series at some point, but I couldn't since the server was crap.

The server was based on outdated version of Hapi, it used an awkward way to register plugins, which I hated, and which they changed a lot in the new version, and I couldn't upgrade easily, plus some 3rd party plugins were taking forever to upgrade to the new Hapi version.

Because of all that, I decided to rewrite and I wanted to see how far I can get without a framework. I ended up with a fully functional backwards-compatible app.

Lessons learned:
- routing is actually a huge part of a framework, at least in my implementation. my "framework" is actually three files in `util`: `chain.js`, `http.js` and `routes.js`
- nodejs http server and middlewares are not as frightening when you get to use them and know a bit of functional programming
- rxjs sucks in simple REST scenarios. Not something i didn't know but i got another confirmation. When there are concurrent events or when you wanna flatten or group collections (like in my lambda) then rxjs wins, otherwise use promises.
- callbacks don't suck that bad, not everything has to be promises. again something I already knew